refactor move to a helper
